requirements.txt
langchain
langchain_community
pypdf2
google-genai
langchain-text-splitters
chromadb
sentence-transformers